requirements.txt
streamlit
pdf2image
python-pptx
pillow
poppler-utils
koreanize-matplotlib
